To be tested
This code as of now is incomplete and untested. Testing can only be done once new compiler cum programmer for ATXmega is acquired.
Current compiler can't deal with float, double or long double longer than 4 bytes. We need more precision, for which we need the double to be 8 bytes. So, a new program has to downloaded, and learnt. Then, this code can be tested.